requirements.txt
lxml==4.3.3
selenium==3.141.0
tornado==6.0.3
calmjs.types==1.0.0
Flask==1.0.2
Twisted==19.2.0
tinycss2==1.0.2
urllib3==1.24.3
Scrapy==1.6.0
requests==2.21.0
browsermob_proxy==0.8.0
beautifulsoup4==4.8.0
calmjs==3.4.1
PyYAML==5.1.2
